azure/requirements.txt
azure-functions>=1.18.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
fastapi>=0.104.0
pydantic>=2.5.0
httpx>=0.25.0
python-dotenv>=1.0.0
openai>=1.0.0
asana>=3.2.0
PyGithub>=1.59.0
pydantic-settings>=2.1.0